Improvements to the standard library (plus related bug fixes in compiler)
- [new] Files.Mod
- [new] LICENSE file and correspoding note in library modules
- [fix] unsigned extension from BYTE to integer types
- [fix] support for recursive POINTER types in Parser
- [fix] error message in Parser for missing formal parameter list in function declaration
- [fix] improved checking of POINTER types in Sema
- [new] better debugging output for symbol files
- [fix] bug in symbol importer on reading exno values
- [fix] bug in semantic checking of string comparisons

+++ test/c/Signedness.c
@@ -5,7 +5,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-int cmp(int x, int y) {
+int cmp(int16_t x, uint8_t y) {
     return x > y;
 }
